services/onlineservice/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
PyJWT==2.10.1
sqlparse==0.5.3
django-prometheus==2.4.1; python_version >= '3.10'
redis
